requirements.txt
alembic==0.9.1
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
cryptography==1.8.1
dominate==2.3.1
enum34==1.1.6
Fabric==1.13.1
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
MySQL-python==1.2.5
ntlm-auth==1.0.2
ordereddict==1.1
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-editor==1.0.3
pywinrm==0.2.2
requests==2.13.0
requests-ntlm==1.0.0
six==1.10.0
SQLAlchemy==1.1.8
visitor==0.1.3
Werkzeug==0.12.1
WTForms==2.1
xmltodict==0.10.2
